--- src/physics/fallingsand/chunks.rs
@@ -1,5 +1,0 @@
-mod chunk;
-mod core;
-mod partial_layer;
-pub mod radial_mesh;
-pub mod util;
